pyproject.toml
[project]
name = "ecommerceapi"
version = "0.1.0"
description = ""
authors = [
    { name = "Hanna Kirieieva", email = "kireevaann2006@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.0,<2.0.0)",
    "ruff (>=0.14.11,<0.15.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)",
    "email-validator (>=2.0.0,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "faker (>=40.1.2,<41.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy-utils (>=0.42.1,<0.43.0)"
]
